Add authorization to delete task

+++ src/tasks/tasks.service.ts
@@ -39,8 +39,8 @@ export class TasksService {
         return task;
     }
 
-    async deleteTask(id:number){
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(user: User, id:number){
+        const result = await this.taskRepository.delete({id, userId:user.id});
         if(result.affected === 0){
             throw new NotFoundException(`Task with ID "${id}" not found.`);
         }
